unix shebang line
Your very first python program! At first glance it's not very
exciting, but there's a lot to learn with this minimalist program.
Run it and you'll see that it is indeed a valid Python program!

Don't skip ahead to more advanced programs like "hello world" until
you understand this code first.

There is a name for the first line of this "empty" program. It is
called "the shebang line" and should appear at the top of all your
python executable programs that you intend to invoke from the unix
command line.

Good habit: Avoid hard-coding anything, but most especially avoid
using absolute paths. Where possible use relative paths or some method
to ask your environment for the feature you want. You can ask unix to
search your PATH environment variable for the location of an
executable file, for example:

    $ echo $PATH
    /home/aaron/bin:/usr/local/bin:/usr/bin:/bin

Each path element is separated by a colon, so the following is the
list of paths that the above line represents:

    /home/aaron/bin
    /usr/local/bin
    /usr/bin
    /bin

Note, that order is important and we will discuss that more below with
an example using the `which` to find the location of a program, the
python binary, in my path:

    $ which python
    /usr/bin/python

The which command will search my PATH until it finds a match. If I
also had a python executable in /home/aaron/bin the which command
would stop searching and return that path instead.

It is important to keep this in mind because it leads to more
flexible systems. Flexibility comes with a caveat: sometimes it can
lead to unintended consequences in new environments if you make
assumptions about where binaries will be found.

Some python-related discussion of shebang:

    Why do people write #!/usr/bin/env python on the first line of a Python script?
    http://stackoverflow.com/q/2429511/462302

    What's the difference between these two python shebangs
    http://stackoverflow.com/q/5709616/462302

A more general discussion of /usr/bin/env:

    How does /usr/bin/env know which program to use?
    http://unix.stackexchange.com/q/12736/11052

Wikipedia explanation: http://en.wikipedia.org/wiki/Shebang_(Unix)
explain shebang

Your very first python program! At first glance it's not very
exciting, but there's a lot to learn with this minimalist program.
Run it and you'll see that it is indeed a valid Python program!

Don't skip ahead to more advanced programs like "hello world" until
you understand this code first.

There is a name for the first line of this "empty" program. It is
called "the shebang line" and should appear at the top of all your
python executable programs that you intend to invoke from the unix
command line.

Good habit: Avoid hard-coding anything, but most especially avoid
using absolute paths. Where possible use relative paths or some method
to ask your environment for the feature you want. You can ask unix to
search your PATH environment variable for the location of an
executable file, for example:

    $ echo $PATH
    /home/aaron/bin:/usr/local/bin:/usr/bin:/bin

Each path element is separated by a colon, so the following is the
list of paths that the above line represents:

    /home/aaron/bin
    /usr/local/bin
    /usr/bin
    /bin

Note, that order is important and we will discuss that more below with
an example using the `which` to find the location of a program, the
python binary, in my path:

    $ which python
    /usr/bin/python

The which command will search my PATH until it finds a match. If I
also had a python executable in /home/aaron/bin the which command
would stop searching and return that path instead.

It is important to keep this in mind because it leads to more
flexible systems. Flexibility comes with a caveat: sometimes it can
lead to unintended consequences in new environments if you make
assumptions about where binaries will be found.

Some python-related discussion of shebang:

    Why do people write #!/usr/bin/env python on the first line of a Python script?
    http://stackoverflow.com/q/2429511/462302

    What's the difference between these two python shebangs
    http://stackoverflow.com/q/5709616/462302

A more general discussion of /usr/bin/env:

    How does /usr/bin/env know which program to use?
    http://unix.stackexchange.com/q/12736/11052

Wikipedia explanation: http://en.wikipedia.org/wiki/Shebang_(Unix)